# Speed up db rider tests