--- a/dropwizard-common-test/src/main/java/org/triplea/dropwizard/test/DropwizardServerExtension.java
+++ b/dropwizard-common-test/src/main/java/org/triplea/dropwizard/test/DropwizardServerExtension.java
@@ -13,7 +13,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.jdbi.v3.core.Jdbi;
 import org.jdbi.v3.core.mapper.RowMapperFactory;
 import org.jdbi.v3.sqlobject.SqlObjectPlugin;
-import org.junit.jupiter.api.extension.AfterAllCallback;
+import org.junit.jupiter.api.extension.AfterEachCallback;
 import org.junit.jupiter.api.extension.BeforeAllCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.api.extension.ParameterContext;
@@ -30,7 +30,7 @@ import org.junit.jupiter.api.extension.ParameterResolver;
  */
 @Slf4j
 public abstract class DropwizardServerExtension<C extends Configuration>
-    implements BeforeAllCallback, AfterAllCallback, ParameterResolver {
+    implements BeforeAllCallback, AfterEachCallback, ParameterResolver {
 
   private static Jdbi jdbi;
   private static URI serverUri;
@@ -60,25 +60,24 @@ public abstract class DropwizardServerExtension<C extends Configuration>
   @Override
   public void beforeAll(final ExtensionContext context) {
     final DropwizardTestSupport<C> support = getSupport();
-    try {
-      log.info("Starting local server for testing..");
-      support.before();
+    log.info("Starting local server for testing..");
+    support.before();
+
+    if (jdbi == null) {
       jdbi =
           Jdbi.create(getDatabase().getUrl(), getDatabase().getUser(), getDatabase().getPassword());
       jdbi.installPlugin(new SqlObjectPlugin());
       rowMappers().forEach(jdbi::registerRowMapper);
-
       log.info("Created JDBI connection to: {}", getDatabase().getUrl());
-      final String localUri = "http://localhost:" + support.getLocalPort();
-      serverUri = URI.create(localUri);
-      log.info("Local server URL set to: {}", localUri);
-    } catch (final RuntimeException e) {
-      log.warn("Ignoring setup error, server already started: {}", e.getMessage());
     }
+
+    final String localUri = "http://localhost:" + support.getLocalPort();
+    serverUri = URI.create(localUri);
+    log.info("Local server URL set to: {}", localUri);
   }
 
   @Override
-  public void afterAll(final ExtensionContext context) throws Exception {
+  public void afterEach(final ExtensionContext context) throws Exception {
     final URL cleanupFileUrl = getClass().getClassLoader().getResource("db-cleanup.sql");
     if (cleanupFileUrl != null) {
       log.info("Running database cleanup..");