--- a/lobby-server/src/test/java/org/triplea/db/dao/api/key/GameHostingApiKeyDaoTest.java
+++ b/lobby-server/src/test/java/org/triplea/db/dao/api/key/GameHostingApiKeyDaoTest.java
@@ -15,19 +15,19 @@ class GameHostingApiKeyDaoTest extends LobbyServerTest {
   private final GameHostingApiKeyDao gameHostApiKeyDao;
 
   @Test
-  @DataSet("game_hosting_api_key/key_exists.yml")
+  @DataSet(value = "game_hosting_api_key/key_exists.yml", useSequenceFiltering = false)
   void keyExists() {
     assertThat(gameHostApiKeyDao.keyExists("game-hosting-key"), is(true));
   }
 
   @Test
-  @DataSet("game_hosting_api_key/key_exists.yml")
+  @DataSet(value = "game_hosting_api_key/key_exists.yml", useSequenceFiltering = false)
   void keyDoesNotExist() {
     assertThat(gameHostApiKeyDao.keyExists("DNE"), is(false));
   }
 
   @Test
-  @DataSet("game_hosting_api_key/insert_key_before.yml")
+  @DataSet(value = "game_hosting_api_key/insert_key_before.yml", useSequenceFiltering = false)
   @ExpectedDataSet("game_hosting_api_key/insert_key_after.yml")
   void insertKey() {
     gameHostApiKeyDao.insertKey("game-hosting-api-key", "127.0.0.2");