--- a/lobby-server/src/test/java/org/triplea/db/dao/moderator/ModeratorAuditHistoryDaoTest.java
+++ b/lobby-server/src/test/java/org/triplea/db/dao/moderator/ModeratorAuditHistoryDaoTest.java
@@ -15,6 +15,12 @@ import org.junit.jupiter.api.Test;
 import org.triplea.modules.http.LobbyServerTest;
 
 @RequiredArgsConstructor
+@DataSet(
+    value =
+        "moderator_audit/user_role.yml,"
+            + "moderator_audit/lobby_user.yml,"
+            + "moderator_audit/moderator_action_history.yml",
+    useSequenceFiltering = false)
 class ModeratorAuditHistoryDaoTest extends LobbyServerTest {
 
   private static final int MODERATOR_ID = 900000;
@@ -23,7 +29,6 @@ class ModeratorAuditHistoryDaoTest extends LobbyServerTest {
   private final ModeratorAuditHistoryDao moderatorAuditHistoryDao;
 
   @Test
-  @DataSet("moderator_audit/pre_insert.yml")
   void addAuditRecordThrowsIfModeratorNameNotFound() {
     assertThrows(
         UnableToExecuteStatementException.class,
@@ -37,8 +42,13 @@ class ModeratorAuditHistoryDaoTest extends LobbyServerTest {
   }
 
   @Test
-  @DataSet("moderator_audit/pre_insert.yml")
-  @ExpectedDataSet("moderator_audit/post_insert.yml")
+  @DataSet(
+      value =
+          "moderator_audit/user_role.yml,"
+              + "moderator_audit/lobby_user.yml,"
+              + "moderator_audit/empty_moderator_action_history.yml",
+      useSequenceFiltering = false)
+  @ExpectedDataSet("moderator_audit/moderator_action_history_post_insert.yml")
   void addAuditRecord() {
     moderatorAuditHistoryDao.addAuditRecord(
         ModeratorAuditHistoryDao.AuditArgs.builder()
@@ -49,7 +59,6 @@ class ModeratorAuditHistoryDaoTest extends LobbyServerTest {
   }
 
   @Test
-  @DataSet("moderator_audit/history_select.yml")
   void selectHistory() {
     List<ModeratorAuditHistoryRecord> results = moderatorAuditHistoryDao.lookupHistoryItems(0, 3);