--- a/lobby-server/src/test/resources/datasets/access_log/user_role.yml
+++ b/lobby-server/src/test/resources/datasets/access_log/user_role.yml
@@ -1,0 +1,3 @@
+user_role:
+  - id: 9111
+    name: PLAYER