--- a/lobby-server/src/test/resources/datasets/lobby_api_key/empty_lobby_api_key.yml
+++ b/lobby-server/src/test/resources/datasets/lobby_api_key/empty_lobby_api_key.yml
@@ -1,0 +1,1 @@
+lobby_api_key: