--- a/maps-server/src/test/resources/db-cleanup.sql
+++ b/maps-server/src/test/resources/db-cleanup.sql
@@ -1,0 +1,5 @@
+-- Deletes table data in proper order
+
+delete from map_skin;
+delete from map;
+delete from category;