## Commits 

commit d7fc3653d9bfe27af5fecceccc7873c4924274da

    Update DropWizardExtension: clean after each test & init JDBI once  & fail if server already started
    
    1. Re-creating DB connection takes some time, if we already have a DB
    connection (ie: jdbi != null), then do not re-created it. This is
    a matter of adding an if null check around the JDBI instantiation.
    
    2. Clean DB after each test, this helps avoid FK violations between
    tests within a test class. The execution of the clean-up is quite
    fast, this makes it easier to write valid tests without slowing
    things down noticably.

    3. To improve error handling and error messaging, configure the drop wizard
    extension server startup (starts server when tests execute) to fail if
    the server is already running. Given we no longer manually start a server
    in CI, this is okay.

commit d7fc3653d9bfe27af5fecceccc7873c4924274da

    Speed up Tests: Update all @DataSet to have 'useSequenceFiltering = false'
    
    This is a big boost in performance for DBRider tests that has more than a
    few tables. SequenceFiltering checks the right order in which to insert tables,
    presumably by attempting permutations. By not doing this, we see pretty significant
    test execution speed-ups.
    
    A downside to this udate is that DbRider will typically fail if we define
    multiple tables in a single YML. The order of tables in YML does not determine
    the order of creation in DbRider. To fix this, we can put each database table
    in its own YML and include multiple YML files per @Dataset. The order of included
    YMLs is deterministic relative to table creation and in this way we can
    ensure tables are created in the right order.

## Additional Notes to Reviewer

**Benchmarks**

Running all tests of the server subprojects in IDEA:

maps_server:  1.8s ->  0.9ms
lobby_server : 124s -> **9.7s**


Running integration test in Travis:  7m 14s ->  **3min 44s**

<!-- Add any additional details that would be helpful to reviewers -->

## Release Note

<!--
Include a release note if there is a bug fix or a visible change for players.
For format & syntax help, see:
https://github.com/triplea-game/triplea/wiki/PR-Release-Notes
-->

<!--RELEASE_NOTE--><!--END_RELEASE_NOTE-->