@DanVanAtta 2020-08-30
@RoiEXLab codacy is interacting through my user; removing the codacy check pattern it went ahead and deleted the initial comment. I was wondering the same and if I had a cat walk across the keyboard or something  :grin: