@DanVanAtta 2020-08-30
(It's actually a TODO on my plate to re-create the codacy link using the service account user and not mine).